Refactoring: remove unused code
#113
