--- data/src/main/java/by/alexandr7035/data/datasource/cache/secrets/SecretsStorage.kt
@@ -1,7 +1,0 @@
-package by.alexandr7035.data.datasource.cache.secrets
-
-interface SecretsStorage {
-    fun saveAccessToken(accessToken: String?)
-
-    fun getAccessToken(): String?
-}
