+++ data/src/main/java/by/alexandr7035/data/helpers/profile_avatars/DicebearAvatarsHelper.kt
@@ -1,7 +1,5 @@
 package by.alexandr7035.data.helpers.profile_avatars
 
-import by.alexandr7035.data.model.local.profile.DicebearImageType
-
 interface DicebearAvatarsHelper {
     fun getImageUrl(
         // See all the possible types on the site
