+++ data/src/main/java/by/alexandr7035/data/helpers/profile_avatars/DicebearImageType.kt
@@ -1,4 +1,4 @@
-package by.alexandr7035.data.model.local.profile
+package by.alexandr7035.data.helpers.profile_avatars
 
 enum class DicebearImageType {
     AVATAR_JDENTICON,
