+++ data/src/main/java/by/alexandr7035/data/repository/LoginRepositoryImpl.kt
@@ -6,8 +6,6 @@ import by.alexandr7035.affinidi_id.domain.model.login.AuthCredentials
 import by.alexandr7035.affinidi_id.domain.model.login.LogOutModel
 import by.alexandr7035.affinidi_id.domain.repository.AppSettings
 import by.alexandr7035.affinidi_id.domain.repository.LoginRepository
-import by.alexandr7035.data.datasource.cache.credentials.CredentialsCacheDataSource
-import by.alexandr7035.data.datasource.cache.secrets.SecretsStorage
 import by.alexandr7035.data.datasource.cloud.ApiCallHelper
 import by.alexandr7035.data.datasource.cloud.ApiCallWrapper
 import by.alexandr7035.data.datasource.cloud.api.UserApiService
@@ -17,8 +15,6 @@ import javax.inject.Inject
 class LoginRepositoryImpl @Inject constructor(
     private val userApiService: UserApiService,
     private val apiCallHelper: ApiCallHelper,
-    private val secretsStorage: SecretsStorage,
-    private val credentialsCacheDataSource: CredentialsCacheDataSource,
     private val appSettings: AppSettings
 ) : LoginRepository {
 
@@ -66,8 +62,9 @@ class LoginRepositoryImpl @Inject constructor(
 
         return when (res) {
             is ApiCallWrapper.Success -> {
-                secretsStorage.saveAccessToken(null)
-                credentialsCacheDataSource.clearCredentialsCache()
+                // TODO clear use case
+//                secretsStorage.saveAccessToken(null)
+//                credentialsCacheDataSource.clearCredentialsCache()
                 LogOutModel.Success
             }
             is ApiCallWrapper.HttpError -> {
