--- domain/src/main/java/by/alexandr7035/affinidi_id/domain/model/profile/SaveProfileModel.kt
@@ -1,6 +1,0 @@
-package by.alexandr7035.affinidi_id.domain.model.profile
-
-data class SaveProfileModel(
-    val userName: String,
-    val userDid: String
-)
